Calculate dimensions from PCB and margins
Add parameters for PCB size and margins on all sides.
Generate the box dimensions from these sizes.
Place the PCB appropriately based on the margins.

This has a change in definition, as the previous PCB size was the size
of the square made by the mounting holes. Now it is the size of the
actual PCB.

Holes are still mounted 5mm inside the edges. Hole positioning will come
soon.