Center the interior box
The interior box (fins for holding the panel) was oddly not centered.
Shift it using the proper parameters for it to be centered.
It also was shifted vertically by 0.1 to remove "the artifact" I'm not
sure what this artifact is, so I'm removing the shift until I see reason
not to. It also makes the interior hole of the feet shorter by 0.1 and
the tolerance under the board 0.1 less, so it seems a bit wrong to have
in place.