Stop translating text by 0.5 and back again.
I suspect this was an artifact before text was linear_extruded, as it
has a default thickness of 1, so shifting by 0.5 placed it on the
surface. Linear extrude was added, and then the calls to LText and CText
were shifted the other direction by 0.5 to get them on the surface. This
now just results in two unnecessary translates, so I've removed them.

I also removed the shift from the "PCB" text. It's thickness has no
bearing on the print, so there's no reason to shift it. If you wanted to
later linear_extrude it, this would get in your way.