Re-work of vent/decoration holes
Make all vent holes Thick + Filet thick so they extend to the end of the
rounded corner, particularly for large filets, but not too far in for
small filets.

Make all decorations Thick/2 so they scale with the thickness.

Separate the vents by a square Thick x Thick pillar.

Generate the position of the first vent/decoration based on the panel
thickness, wall thickness, margin, and vent spacing so that there is one
pillar's width beteween the interior edge and the first vent.

Bring the decoration/vent in from 1mm outside the box both horizontally
and vertically to prevent artifacts. This does result in vents reaching
under the box farther than previously, but I think the aesthetic is
pleasing, and it handles strange cases better, such as thick walls and
tiny filets.