Make it clear Filet is a radius
Update the comment at the top, which described Filet as a diameter,
which was incorrect for most of the file.

Make it clear that the circle used in foot creation is a radius. It
already was, but provide r= in the argument so it's unambiguous.

SquareHole was treating its filet as a diameter, rather than a radius.
Switch it to treat Filet as a radius. This does make the example square
holes rounder.