Parametrize the box fixation tabs
This is the last part to make the box fully scalable. Now if you double
every parameter, the box doubles in all dimensions. The panel features
are not changed, however. That's up to you to fix.

The tab cutouts are now scaled off of the screw hole size, since that
fixes the diameter of the tab.

I've also added an OuterMargin which adjusts the spacing you need
between the tab and the box wall. This is a property of your printer,
and shouldn't be changed when you scale other aspects of the box. I
might base other dimensions, like the panel gap on this value as well.