Scale interior filet to keep concentric corners
If the filet is more than double the thickness, scale the interior
filets so that the walls have a constant thickness around the corners
and the interior and exterior filets have a common center axis. If the
filets are smaller, then keep the old behavior, making the corners
mostly square.