Support arbitrary legs on top/bottom shell
Add BLLeg, FLLeg, BRLeg, and FRLeg parameters where 0 means put that leg
on the bottom shell, and 1 means put it on the top. Call the legs and
hols modules with a "top" argument where 0 means build the bottom shell
legs/holes, and 1 means build the top. Gate each leg or hole creation
based on logic between XXLeg and top.

This also required making the top shell using mirror() rather than
rotate, since that kept the same leg/hole in the same position when
flipped to the top.