Support experimental snap tabs
Set SnapTabs to 1 and get tabs with buttons and semispherical holes in
the box. This is experimental because I haven't built a box like this
and don't know how well it will work. The thickness of the tab is
probably important since you need it to flex just the right amount to
allow the button to fit. Also, the ScrewHole value is used for the size
of the button, so it should not exceed 2*Thick.