Move panel management to the top
Break out the panel holes and text into separate modules.

Put those modules at the top of the file so they are easier to find.
This unfortunately puts them before the modules they instantiate, but I
think this is a fair trade-off for readability and findability of things
that should be customized.