Add experimental screwless PCB feet
If Screwless is true, use feet with a peg in them that goes through the
PCB holes. Feet then come down from the top shell to press against the
PCB and hold it in place. These have a hole in the center so the peg
from the bottom shell will insert into it and keep the shells in a
stable position in the X-Y plane. If you combine this with screwless
snap tabs, you can have a fully screwless box.