Think of panel gaps as on a single edge.
Panel gaps were previously for both edges, as in the reduction in size
of the panel. The code often divides this value by 2 to think about a
single edge. It makes more sense to just store this value as on a single
edge and just double it in the cases we need it, which is fewer than the
cases were we previously cut it in half. This is fewer opterations and
easier to reason about.